#bugfix - Parametric crash from Paul 8/20/15 - caused when using the "Add" variable in quick setup to remove variables by unchecking. Workaround is to use the "Remove" button in the parametric quick setup to remove variables. #bugfix - debug assert errors from wxWidgets fixed by type casting size_t to int in wxString::Format statements in graph.cpp and simulation.cpp

+++ src/graph.cpp
@@ -451,7 +451,7 @@ void GraphCtrl::Display(std::vector<Simulation *>sims, Graph &gi)
 			if (count > 0)
 			{
 				yvars.push_back(vv);
-				ynames.push_back(sims[i]->GetLabel(m_g.Y[0]) + wxString::Format(" : run %d", i+1));
+				ynames.push_back(sims[i]->GetLabel(m_g.Y[0]) + wxString::Format(" : run %d", (int)(i+1)));
 			}
 		}
 	}
@@ -537,7 +537,7 @@ void GraphCtrl::Display(std::vector<Simulation *>sims, Graph &gi)
 		// single value axis
 		wxPLLabelAxis *x1 = new wxPLLabelAxis(-1, yvars.size(), m_g.XLabel);
 		for (size_t i = 0; i<ynames.size(); i++)
-			x1->Add(i, wxString::Format("%d", i+1));
+			x1->Add(i, wxString::Format("%d", (int)( i+1)));
 //			x1->Add(i, ynames[i]);
 		SetXAxis1(x1);
 	}
